Refactor application package
- Remove infrastructure dependency (even though it was test dependency)

Refactor integration test into unit test
