+++ packages/application/src/use-case/index.ts
@@ -1,1 +1,2 @@
+export * from "./errors";
 export * from "./registerUser";
