+++ packages/application/tsconfig.json
@@ -8,9 +8,6 @@
   "references": [
     {
       "path": "../core"
-    },
-    {
-      "path": "../infrastructure"
     }
   ]
 }
